Move dividends to its own handler
